+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -2,9 +2,9 @@ package wooteco.subway.path.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.fare.domain.AgeFarePolicy;
 import wooteco.subway.line.application.LineService;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.line.domain.fare.AgeFarePolicy;
 import wooteco.subway.member.domain.User;
 import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.path.dto.PathResponse;
