+++ src/main/java/wooteco/subway/path/dto/PathResponseAssembler.java
@@ -1,6 +1,6 @@
 package wooteco.subway.path.dto;
 
-import wooteco.subway.line.domain.fare.AgeFarePolicy;
+import wooteco.subway.fare.domain.AgeFarePolicy;
 import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.station.dto.StationResponse;
